+++ NEWS.md
@@ -2,6 +2,7 @@
   * bug fixes
   * additional tests added
   * sar_pred() function added for SAR extrapolation
+  * negative fitted value check (fit$neg_check) returned for each individual model fit
   
 ## Version 1.1.1
  * added vignette